+++ .gitignore
@@ -17,3 +17,4 @@
 
 /frontend
 /backend
+/client/*
