--- client/web/footer.html
@@ -1,3 +1,0 @@
-<footer>
-
-</footer>
