--- client/web/homePage.html
@@ -1,30 +1,0 @@
-<!doctype html>
-<html lang="en">
-    {{template "head.html" .}}
-<body>
-    <h1>{{.Title}}</h1>
-    <form action="/" method="post">
-        <label for="fullLink">Full link</label>
-        <input type="url" name="fullLink" id="fullLink">
-        <button type="submit" name="submint">Generate</button>
-    </form>
-    {{ if ne .Error "" }}
-    <p>{{.Error}}</p>
-    {{else}}
-    <div class="d-flex flex-column h-100">
-        <table class="table">
-            <th>Исходная ссылка</th>
-            <th>Короткая ссылка</th>
-            <th>Ссылка для просмотра статистики</th>
-            <th>Дата и время создания короткой ссылки</th>
-            <tr>
-                <td>{{.FullLink}}</td>
-                <td>{{.ShortLink}}</td>
-                <td>{{.StatLink}}</td>
-                <td>{{.CreatedAt | dateFormat "02-01-2006 15:04:05"}}</td>
-            </tr>
-        </table>
-    </div>
-    {{end}}
-</body>
-</html>
