--- client/web/statPage.html
@@ -1,24 +1,0 @@
-<!doctype html>
-<html lang="en">
-{{template "head.html" .}}
-<body>
-    <h1>{{.Title}}</h1>
-<h3>Статистика переходов по короткой ссылке {{.ShortLink}}</h3>
-    <h4>Дата создания {{.CreatedAt | dateFormat "02-01-2006 15:04:05"}}&nbspОбщее количество переходов&nbsp{{.TotalCount}}</h4>
-<div class="d-flex flex-column h-200">
-  <table class="table">
-    <caption>Количество переходов с разбивкой по IP адресам</caption>
-    <th>IP address</th>
-    <th>Количество переходов</th>
-    <th>Дата и время последнего перехода</th>
-    {{range .FollowList}}
-    <tr>
-      <td>{{.IPaddress}}</td>
-      <td>{{.Count}}</td>
-      <td>{{.FollowLinkAt | dateFormat "02-01-2006 15:04:05"}}</td>
-    </tr>
-    {{end}}
-  </table>
-</div>
-</body>
-</html>
