+++ internal/api/handler/router.go
@@ -1,7 +1,7 @@
 package handler
 //
 //import (
-//	"github.com/a-ivlev/URL-shortener/shortener/internal/app/redirectBL"
+//	"github.com/a-ivlev/URL-shortener/internal/app/redirectBL"
 //	"context"
 //	"encoding/json"
 //	"fmt"
