+++ internal/app/repository/followingBL/followingBL.go
@@ -1,9 +1,9 @@
 package followingBL
 
 import (
-	"github.com/a-ivlev/URL-shortener/shortener/internal/app/repository/shortenerBL"
 	"context"
 	"fmt"
+	"github.com/a-ivlev/URL-shortener/internal/app/repository/shortenerBL"
 	"log"
 	"time"
