+++ internal/db/inmemoryDB/shortenerMapDB.go
@@ -1,10 +1,10 @@
 package inmemoryDB
 
 import (
-	"github.com/a-ivlev/URL-shortener/shortener/internal/app/repository/shortenerBL"
 	"context"
 	"database/sql"
 	"errors"
+	"github.com/a-ivlev/URL-shortener/internal/app/repository/shortenerBL"
 	"github.com/google/uuid"
 	"log"
 	"strings"
